portfolio/VolgaCTF-2026/stego/deep/task/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch>=2.6
torchvision
tqdm
Pillow
reedsolo
